// src/pages/Report/models/models.ts
export type hisOrder = {
  ABACKUPDATE: any;
  ABUYSELL: string;
  ACLIENTCODE: string;
  ACONTRACTNO: null;
  ADATETIME: any;
  AERRORCODE: number;
  AEXCHANGE: string;
  AEXCHANGEORDERNO: string;
  ALASTTRADEDTIME: string;
  AMARKETSEGMENT: string;
  AMESSAGE_EN: string;
  AMESSAGE_VN: string;
  AMODIFYQTY: number;
  AORDERID: number;
  AORDERSESSION: string;
  AORDERSTATUS: string;
  AORDERSTATUS_VN: string;
  AORDERTYPE: string;
  AORDERTYPE_VN: string;
  APLACEDBY: string;
  APRICE: number;
  APRICETYPE: string;
  APRODUCTTYPE: string;
  AQUANTITY: number;
  AREMAININGQTY: number;
  ASTOCKCODE: string;
  ATRADEDQTY: 100;
  ATRADINGACCOUNT: string;
  ATRADINGACCOUNT_VN: string;
  ATRANID: number;
};
export type ReportData = {
  ACHARGE: number;
  ACHARGE3: number;
  ACHARGE_DTV: number;
  ACLIENTCODE: string;
  AEXCHANGE: string;
  AMATCHED_AMT: number;
  AMATCHED_PRC: number;
  AMATCHED_QTY: number;
  AORDERDATE: Date;
  AORDERNO: string;
  AORDERTYPE: string;
  AORGORDER_AMT: number;
  AORGORDER_PRC: string;
  AORGORDER_QTY: string;
  APTYID: number;
  ARMN_QTY: number;
  ASTOCKCODE: string;
  ATRADEDATE: Date;
  ATRADENO: string;
};
export type hometransferds = {
  ACLIENTCODE: string;
  ACASHDEPOSITED: number;
  ALEDGERBALANCE: number;
  ACASHADVANCE: number;
  AADHOCCASH: number;
  AMARGINPRO: number;
  ACASHTRADING: number;
  ACASHMARGIN: number;
  ACASHUSED: number;
  ACASHFEE: number;
  ACASHTRANSFER: number;
  ADEBT: number;
  AFSAVING: number;
  TIENMAT: number;
  AVAIL_TRADINGCASH: number;
  AVAIL_STOCKVAL: number;
  AVAIL_TRANSFER: number;
  REMAININGQUOTA: number;
  AVAIL_PLK: number;
  AVAIL_FSAVING: number;
  AMARPRODEBT: number;
  ACASHHOLDINGBUY: number;
};